Move notice.amendments to notice.amendments.fetch
